build(yaml): remove additional truthy values
Remove additional truthy values of "yes" or "no", because when Ansible Lint runs in AstroNvim, it does not
read from .yamllint in the project root. When running Ansible Lint on the command line, it does take
.yamllint into account. Nobody has time to debug why that is happening. Another good reason, is that the
Ansible documentation uses "true" or "false", so might as well stick with that. It saves on the effort of
adding the additional values to a .yamllint with every project.